move `method` to webfinger class
